added integration tests and updated readme

+++ hook/config.go
@@ -6,7 +6,7 @@ type Config struct {
 	// Host defines the hostname of the fluent service
 	Host string
 	// Port defines the port the fluent service is listening on
-	Port uint64
+	Port uint16
 	// Base Tag defines the tag which will appear in fluent
 	BaseTag string
 	// Levels defines the levels for which the hook will be fired
